+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update some color usage
   - SONAR-12539 Update leak color of timeline
   - SONAR-12541 Rewrite GlobalMessages to @emotion and clean colors and css
+  - SONAR-11601 More colorblind friendly coverage indicators
 
 ## 0.0.31