+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.0.32
+
 - Fix BoxedTabs style
 - Update some color usage
   - SONAR-12539 Update leak color of timeline